Introduced multi-processing
classicPSOswarm.py & clanPSOswarm.py
------------------------------------------------------------
1) Rearranged the order of some inputs
2) Renamed "convex_boundaries" to "spawn_boundaries", so that the field name gives more information about its functionality and to separate it from search boundaries.

particle.py
----------------
The particle velocity is initialized with the "zeros" numpy function, which creates the array/vector faster than by manually creating it with a for loop.

benchmark_functions.py
------------------------------------
All benchmark functions are now called from this script instead of being written in the "main.py" file. This is for readability, maintainability and expandability puproses.

experiment.py
---------------------
Runs ONE experiment. By the word "experiment", a full execution (NOT iteration) of the PSO algorithm is meant. 

experiment_data_creation.py
------------------------------------------
Executes a specified number of experiments (see "experiment.py" above for the definition of the word "experiment". Multi-processing is introduced here (each experiment is calculated in a different processor).