Executed a few functions to show previous commit changes nothing.
Adding data files in "Experiment with maximum implementation" with some of the benchmark functions. More specifically, experiments were conducted on the sphere function, the quadric function and the rastrigin function. This was done to attest to the fact that the (insignificant) change of decentralizing the benchmark function characteristics in a different file, as well as other changes for readability reasons (see commit "8f336c0b4512e96cba839573772d4a3463f5c4a4") had no impact on the execution of the program. The experimental results are at the same scale (albeit with little numeric differences due to the embedded stochasity of the algorithm)