Maximization implementation to minimization one. No performance changes.
With this commit, some example experiments were conducted. No significant change in performance is detected. 
Worthy to note that the files were cross-referenced with the ones contained in the "Experiments with maximum implementation", which in turn have been checked to have similar performance to the ones originally conduted "final experiments".