Created new data generation pipeline.
This commit aims to improve the readability of the code chunks which produce data by making them independent of the "main.py" module.
For this purpose, three new modules were created:
1) experiment2.py
2) experiments_data_creation2.py
3) experimental_data_manipulation.py
Note: name of the files are subject to change.

The abstraction of the parameter names being encapsulated inside a "kwargs" dictionary object -- which is unpacked when the arguments require to be passed into the "PSO" classes-- aims at higher code readability and maintainability.
An example for the latter is the addition of new PSO variations, that may require different parameters/factors for their execution. The new parameters may be inserted inside the "kwargs" variable and adjusted in the "experiment2.experiment" function (as things are now).

main.py
-----------
The "main" function has been altered in a way to showcase how easier and readable the creation of data has become with this new installment.

"benchmark_functions.py"
--------------------------------------
Added a 'name' key-value pair to each different function. Currently, this was implemented so that the "main" function of "main.py" may create the file names in a more abstract way: just by looking at the "selected" function ("main.py" -> line 36).


All other modules
--------------------------
Negligible changes.