Read data from external file, moved main.py, module paths corrections.
==================
|| MODIFICATIONS ||
==================

clan_pso.py
-----------------
`conference_pso_behaviour` may now receive a class object directly, instead of having to provide an iterable of classes. Currently NOT being utilized.

main.py
------------
Moved the main file one hierarchical level up, so as to make relative importing easier to manage.
Expects a JSON file as a console argument to acquire the PSO data initial configuration.

==============
|| ADDITIONS ||
==============
"configs" folder
-----------------------
A folder which contains two JSON files: one for a classic PSO and one for a clan PSO. The purpose of this is to be able to keep the source code identical, but easily designing new experiment parameterization.

 extractor.py
-------------------
A module whose purpose is to log configurations of the swarm.
Currently, the positions of a single configuration of either "classic" or "clan" instance is logged in a pandas.DataFrame. This module wil most probably be developed further.
Results of this module may be stored in any file format (e.g. csv files). The final goal is to analyze the data and extract useful experimental conclusions.

config_initialization_handler.py
---------------------------------------------
Module which can handle both "classic" and "clan" PSO input data. Currently, there are a few entries which might be unnecessarily repeated (e.g. "objective_function"). An argument for allowing this is the fact that this approach offers complete generality to the user, in case they wish to conduct experiments with potentially unintuitive setups (e.g. different objective functions per clan).