Readable "main.py", "experiment.py".Experiments are logged. Read comment
==============
|| DELETIONS ||
==============
Removed all legacy code, which was not compatable with the new architecture of the codebase.

==================
|| MODIFICATIONS ||
==================

main.py
------------
1) Deleted legacy code fo main.py, which was much less readable, as well as much less maintainble.
This was possible thanks to the change in perspective on how the parameterization of the experiment(s) is inserted in the program: formerly, it was hard-coded inside "main.py", while now the configuration is done via an (external) JSON file.
2) The user is able to choose the experiments being conducted in parallel (which would be the CPU-intensive use) or not (in cases where the user wishes to use their CPU in other tasks, perhaps).

experiment.py
---------------------
The "run()" function has been reworked from scratch, so to allow for a much more readable execution loop.

extractor.py
------------------
The "gbest_extractor" function has been added.


==============
|| ADDITIONS ||
==============

logger.py
--------------
A module which automatically creates directories for storing conducted experiments.
The paths generated respect the experimentation parameterization and split the directories according to it (up to an extent). In order to discern the experiments with absolute certainty, the input configuration file is copied inside the generated folder.
Concurrent writing has been handled with the use of a (single) semaphore when generating paths, so no bugs with regards to that should be observed.