Fix:Clan PSO; `current iteration` -> `conference_kwargs`. Read comment.
The current_iteration of the algorithm is now passed into the `conference_pso_kwargs` such that the StandardPSO (which is very oftenly used PSO variation) is applied correctly.
Reminder: the `current_iteration` is responsible for decrementing the inertia weight ω of the particles.

WARNING
---------------
The current implementation is hard-coded; it explicitly attaches the value of the last clan to be iterated/examined by the for loop that has the `_current_iteration` attribute.
NOTE: Here, there is an assumption that all clans are synchronized and, thus, their `_current_iteration` attribute (should they have it) must be identical.