Fix: Removed unnecessary `if` conditions: the vls r clamped in interval
The previous implementation was slightly different than the one described in the paper (although the effects of this change is suspected to not cause an observable change).
The previous implementation disallowed the coefficients to drop below or exceed the coefficient limits. This means that (depending on the choice of bounds and initial coefficient values), the coefficients might never become the boundaries.
On the other hand, the current implementations allows it. Of course, after the limits are exceeded the results are clampled.

TLDR: In previous implementation, clamping might have been useless, while this has now been fixed.