Fix: IMPORTANT! Concurrency experiments overwritting folders solved.Cmnt
1) Logical mistake:
Previously, two different processes could read the same (maximum) `experiment_id` currently in the folder, which would be the (identical) base for defining the next folder name. Now, the folder is created first BEFORE a different process has the opportunity to read the maximum folder name. Therefore, the folder created by the previous process is picked.
2) Replaced to correct data structure: `threading.BoundedSemaphore` to `multiprocessing.BoundedSemaphore`. This is because the implementation (see `main.py`) creates a new process for each experiment and not  a thread (`ProcessPoolExecutor`).
3) In cases where a folder is to be copied, an error is raised by the `os.makedirs` command, which was ignored previously.