READABIL:  Removed `_function` postfix for clarity
Admittedly, this change in code was not super important. It mostly serves to remove the `+ '_function'` part of the `getattr`, which would have future developers wondering why that would be necessary without diving deeper into the source code.

Moreover, understanding of the keyword `sphere`, `quadric` etc. should remain the same, as their namespace (i.e. `benchmark_functions` should make it clear that the variables refer to functions.