UX: small changes to quick webapp that plots means
Made some small changes to make the GUI a bit more understandable for first-time users.

Although this is not strictly part of the framework we wish to implement, I provide a prototype script in case users wish to quickly visualize the convergence of their algorithm fast.

This is achieved by a small web-app that:
1) reads paths that are
	a) inserted by the user and
	b) associated with a keyword, which denotes the (sub)name of the filename.
2) averages all experiements files collected, and.
3) displayes them with a line plot to the user.

The user then has the ability to filter which results are being displayed to them, like:
i) displaying values only withing a user-inserted range of values.
ii) displaying values only withing a user-inserted range of iterations.

The web-app can plot multiple traces one by one.

As this serves as a quick verification tool. This feature is not not planned to be expanded further and users are encouraged to either find other existing plotting packages or implement their own tools.